migration

+++ database/migrations/2023_09_04_082432_create_project_items.php
@@ -16,7 +16,7 @@ class CreateProjectItems extends Migration
         Schema::create('project_items', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('project_id')->unsigned()->nullable();
-            $table->BigInteger('city_id')->unsigned()->nullable();
+            $table->bigInteger('city_id')->unsigned()->nullable();
             $table->bigInteger('pic_id')->unsigned()->nullable();
             $table->bigInteger('item_id')->unsigned()->nullable();
             $table->bigInteger('vendor_price')->default(0);
